Spashscreen for Timberwols as huge headerfile
This header file contains ~12000 lines of splashscreen.
Move it to a separate commit to make the other changes easier to read.

I do not claim any authorship of this code.

This is merely a refactor of the code drop at
https://github.com/tfrieden/timberwolf
which is a project to port the firefox 4.0.1 browser to Amiga OS4.

The upstream branch that tfrieden's port is based on is:
https://github.com/mozilla/gecko-dev/tree/GECKO50_2011051718_RELBRANCH

This repository restores the full commit history of firefox 4.0.1 which
tfrieden's port is built upon.